Update to kubebuilder v3 structure
