+++ apis/sample/v1alpha1/doc.go
@@ -14,8 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package v1alpha1 contains the v1alpha1 group Sample resources of the Template provider.
-// +kubebuilder:object:generate=true
-// +groupName=sample.template.crossplane.io
-// +versionName=v1alpha1
 package v1alpha1
