+++ apis/v1alpha1/doc.go
@@ -14,8 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package v1alpha1 contains the core resources of the Template provider.
-// +kubebuilder:object:generate=true
-// +groupName=template.crossplane.io
-// +versionName=v1alpha1
 package v1alpha1
