+++ go.mod
@@ -1,9 +1,9 @@
 module github.com/crossplane/provider-template
 
-go 1.17
+go 1.16
 
 require (
-	github.com/crossplane/crossplane-runtime v0.14.1-0.20210812020058-ba474e81c62c
+	github.com/crossplane/crossplane-runtime v0.14.0
 	github.com/crossplane/crossplane-tools v0.0.0-20210320162312-1baca298c527
 	github.com/google/go-cmp v0.5.6
 	github.com/pkg/errors v0.9.1
